use intellij to delegate methods
