+++ azkaban-common/src/test/java/azkaban/executor/JdbcExecutorLoaderTest.java
@@ -1060,7 +1060,8 @@ public class JdbcExecutorLoaderTest {
 
     //TODO kunkun-tang: temporary work-around here. This Test is to be deprecated.
     return new JdbcExecutorLoader(props,
-        new CommonMetrics(new MetricsManager(new MetricRegistry())), null, null);
+        new CommonMetrics(new MetricsManager(new MetricRegistry())), null
+        , null, null);
   }
 
   private boolean isTestSetup() {
